Fix broken local append
In `local_patch_append`, when `source` and `target` are the same, the append command will fail. By redirect the output of cat into the source file, the source file gets cleared immediately before the command proper gets executed, resulting in only the append file's contents being written to the `target` (i.e. also `source`). Thus, when they are the same, we cat only the append file and redirect it into the source file with append mode.